refactor(checkout): CHECKOUT-2951 Move `createCheckoutClient`, `createCheckoutService` and `createCheckoutStore` to checkout module
