+++ src/core/checkout/create-checkout-service.js
@@ -1,17 +1,17 @@
 import { createRequestSender } from '@bigcommerce/request-sender';
 import { createClient as createPaymentClient } from 'bigpay-client';
-import { BillingAddressActionCreator } from './billing';
-import { CartActionCreator } from './cart';
-import { CheckoutService } from './checkout';
-import { ConfigActionCreator } from './config';
-import { CountryActionCreator } from './geography';
-import { CouponActionCreator, GiftCertificateActionCreator } from './coupon';
-import { createCustomerStrategyRegistry, CustomerActionCreator } from './customer';
-import { OrderActionCreator } from './order';
-import { createPaymentStrategyRegistry, PaymentMethodActionCreator } from './payment';
-import { InstrumentActionCreator, InstrumentRequestSender } from './payment/instrument';
-import { QuoteActionCreator } from './quote';
-import { createShippingStrategyRegistry, ShippingCountryActionCreator, ShippingOptionActionCreator } from './shipping';
+import { BillingAddressActionCreator } from '../billing';
+import { CartActionCreator } from '../cart';
+import { CheckoutService } from '../checkout';
+import { ConfigActionCreator } from '../config';
+import { CountryActionCreator } from '../geography';
+import { CouponActionCreator, GiftCertificateActionCreator } from '../coupon';
+import { createCustomerStrategyRegistry, CustomerActionCreator } from '../customer';
+import { OrderActionCreator } from '../order';
+import { createPaymentStrategyRegistry, PaymentMethodActionCreator } from '../payment';
+import { InstrumentActionCreator, InstrumentRequestSender } from '../payment/instrument';
+import { QuoteActionCreator } from '../quote';
+import { createShippingStrategyRegistry, ShippingCountryActionCreator, ShippingOptionActionCreator } from '../shipping';
 import createCheckoutClient from './create-checkout-client';
 import createCheckoutStore from './create-checkout-store';
