+++ src/core/common/error/index.js
@@ -1,1 +1,2 @@
+export { default as createRequestErrorFactory } from './create-request-error-factory';
 export { default as RequestErrorFactory } from './request-error-factory';
