+++ src/core/config/index.js
@@ -1,3 +1,4 @@
 export { default as ConfigActionCreator } from './config-action-creator';
 export { default as ConfigSelector } from './config-selector';
 export { default as configReducer } from './config-reducer';
+export { default as ConfigRequestSender } from './config-request-sender';
