+++ src/core/coupon/coupon-action-creator.spec.js
@@ -1,8 +1,8 @@
 import { Observable } from 'rxjs';
+import { createCheckoutStore } from '../checkout';
 import { getCouponResponseBody } from './coupon.mock';
 import { getErrorResponse, getResponse } from '../common/http-request/responses.mock';
 import * as actionTypes from './coupon-action-types';
-import createCheckoutStore from '../create-checkout-store';
 import CouponActionCreator from './coupon-action-creator';
 
 describe('CouponActionCreator', () => {
