+++ src/core/payment/strategies/afterpay-payment-strategy.spec.ts
@@ -2,7 +2,7 @@ import { createScriptLoader } from '@bigcommerce/script-loader';
 import { merge } from 'lodash';
 import { createClient as createPaymentClient } from 'bigpay-client';
 import { CartActionCreator } from '../../cart';
-import { CheckoutStore } from '../../checkout';
+import { createCheckoutClient, createCheckoutStore, CheckoutStore } from '../../checkout';
 import { createPlaceOrderService, OrderRequestBody, PlaceOrderService } from '../../order';
 import { RemoteCheckoutPaymentError, RemoteCheckoutSessionError } from '../../remote-checkout/errors';
 import { createRemoteCheckoutService, RemoteCheckoutService } from '../../remote-checkout';
@@ -13,8 +13,6 @@ import AfterpayPaymentStrategy from './afterpay-payment-strategy';
 import AfterpayScriptLoader from '../../remote-checkout/methods/afterpay';
 import CheckoutClient from '../../checkout/checkout-client';
 import PaymentMethod from '../payment-method';
-import createCheckoutClient from '../../create-checkout-client';
-import createCheckoutStore from '../../create-checkout-store';
 
 describe('AfterpayPaymentStrategy', () => {
     let client: CheckoutClient;
