+++ src/core/payment/strategies/credit-card-payment-strategy.spec.ts
@@ -1,10 +1,8 @@
 import { omit } from 'lodash';
 import { createClient as createPaymentClient } from 'bigpay-client';
-import { CheckoutStore } from '../../checkout';
-import { getOrderRequestBody } from '../../order/orders.mock';
+import { createCheckoutClient, createCheckoutStore, CheckoutStore } from '../../checkout';
 import { createPlaceOrderService, PlaceOrderService } from '../../order';
-import createCheckoutClient from '../../create-checkout-client';
-import createCheckoutStore from '../../create-checkout-store';
+import { getOrderRequestBody } from '../../order/orders.mock';
 import CreditCardPaymentStrategy from './credit-card-payment-strategy';
 
 describe('CreditCardPaymentStrategy', () => {
