+++ src/core/payment/strategies/offsite-payment-strategy.spec.js
@@ -1,8 +1,8 @@
 import { merge, omit } from 'lodash';
+import { createCheckoutStore } from '../../checkout';
 import { getOrderRequestBody, getIncompleteOrder, getSubmittedOrder } from '../../order/orders.mock';
 import { OrderFinalizationNotRequiredError } from '../../order/errors';
 import * as paymentStatusTypes from '../payment-status-types';
-import createCheckoutStore from '../../create-checkout-store';
 import OffsitePaymentStrategy from './offsite-payment-strategy';
 
 describe('OffsitePaymentStrategy', () => {
