+++ src/core/remote-checkout/remote-checkout-service.spec.js
@@ -2,6 +2,7 @@ import { createAction } from '@bigcommerce/data-store';
 import { createRequestSender } from '@bigcommerce/request-sender';
 import { Observable } from 'rxjs';
 import { BillingAddressActionCreator } from '../billing';
+import { createCheckoutClient, createCheckoutStore } from '../checkout';
 import { RemoteCheckoutSynchronizationError } from './errors';
 import { ShippingAddressActionCreator } from '../shipping';
 import { getBillingAddress } from '../billing/billing-address.mock';
@@ -9,8 +10,6 @@ import { getShippingAddress } from '../shipping/shipping-address.mock';
 import * as actionTypes from './remote-checkout-action-types';
 import * as billingActionTypes from '../billing/billing-address-action-types';
 import * as shippingActionTypes from '../shipping/shipping-address-action-types';
-import createCheckoutClient from '../create-checkout-client';
-import createCheckoutStore from '../create-checkout-store';
 import RemoteCheckoutActionCreator from './remote-checkout-action-creator';
 import RemoteCheckoutRequestSender from './remote-checkout-request-sender';
 import RemoteCheckoutService from './remote-checkout-service';
