+++ src/core/shipping/strategies/amazon-pay-shipping-strategy.spec.ts
@@ -3,7 +3,7 @@
 
 import { createScriptLoader } from '@bigcommerce/script-loader';
 import { AmazonPayScriptLoader } from '../../remote-checkout/methods/amazon-pay';
-import { CheckoutStore } from '../../checkout';
+import { createCheckoutClient, createCheckoutStore, CheckoutStore } from '../../checkout';
 import { NotInitializedError } from '../../common/error/errors';
 import { RemoteCheckoutAccountInvalidError, RemoteCheckoutSessionError, RemoteCheckoutShippingError } from '../../remote-checkout/errors';
 import { createRemoteCheckoutService, RemoteCheckoutService } from '../../remote-checkout';
@@ -15,8 +15,6 @@ import { getRemoteCustomer } from '../../customer/customers.mock';
 import { getShippingAddress } from '../shipping-address.mock';
 import AmazonPayShippingStrategy from './amazon-pay-shipping-strategy';
 import UpdateShippingService from '../update-shipping-service';
-import createCheckoutClient from '../../create-checkout-client';
-import createCheckoutStore from '../../create-checkout-store';
 import createUpdateShippingService from '../../shipping/create-update-shipping-service';
 
 describe('AmazonPayShippingStrategy', () => {
